requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# ECM_Implement/main.py: 2
# ECM_Implement/main_notebook.ipynb: 1
convokit == 2.4.3

# ECM_Implement/main_notebook.ipynb: 25
graphviz == 0.15

# ECM_Implement/main.py: 10
# ECM_Implement/main_notebook.ipynb: 11
matplotlib == 3.3.3

# ECM_Implement/main.py: 6,7
# ECM_Implement/main_notebook.ipynb: 6,7
nltk == 3.5

# ECM_Implement/main.py: 9,22
# ECM_Implement/main_notebook.ipynb: 10,23
numpy == 1.18.5

# ECM_Implement/main.py: 3
# ECM_Implement/main_notebook.ipynb: 3
pandas == 1.1.4

# ECM_Implement/main_notebook.ipynb: 25
pydot == 1.4.1

# ECM_Implement/main.py: 11
# ECM_Implement/main_notebook.ipynb: 12
seaborn == 0.11.0

# ECM_Implement/main.py: 21
# ECM_Implement/main_notebook.ipynb: 22
sklearn == 0.0

# ECM_Implement/main.py: 19,20
# ECM_Implement/main_notebook.ipynb: 20,21
tensorflow-gpu == 2.6.0
